:hammer: [redux] 重新编排redux类型。并增加了统一的state用于访问
