+++ src/shared/redux/types/actor.ts
@@ -1,3 +1,5 @@
+import { Record, List } from 'immutable';
+
 export interface ActorType {
   id: number;
   uuid: string;
@@ -15,3 +17,20 @@ export interface ActorType {
 export interface ActorDataType {
   [key: string]: string;
 }
+
+export type ActorState = Record<{
+  isFindingTemplate: boolean;
+  suggestTemplate: List<any>;
+  findingResult: List<any>;
+  currentEditedTemplate: Record<{
+    uuid: string;
+    name: string;
+    desc: string;
+    avatar: string;
+    info: string;
+  }>;
+  selfTemplate: List<any>;
+  selectedTemplate: any;
+  selfActors: List<any>;
+  selectedActorUUID: string;
+}>;
