+++ src/shared/redux/types/cache.ts
@@ -1,0 +1,8 @@
+import { Record, Map } from 'immutable';
+
+export type CacheState = Record<{
+  user: Map<string, any>;
+  template: Map<string, any>;
+  actor: Map<string, any>;
+  group: Map<string, any>;
+}>;
