+++ src/shared/redux/types/group.ts
@@ -1,3 +1,4 @@
+import { Record, Map, List } from 'immutable';
 import { ActorType } from './actor';
 
 export interface GroupActorMsgData {
@@ -17,3 +18,14 @@ export interface GroupActorType {
   actor_info: {};
   actor: ActorType;
 }
+
+export type GroupState = Record<{
+  info: Map<string, any>;
+  invites: List<any>;
+  groups: List<any>;
+  selectedGroupUUID: string;
+  isFindingGroup: boolean;
+  findingResult: List<any>;
+  requestingGroupUUID: List<string>;
+  groupActorMap: Map<string, Map<string, string>>;
+}>;
