+++ src/shared/redux/types/note.ts
@@ -1,0 +1,8 @@
+import { Record, Map } from 'immutable';
+
+export type NoteState = Record<{
+  noteList: Map<string, any>;
+  selectedNoteUUID: string;
+  isSync: boolean;
+  isSyncUUID: string;
+}>;
