+++ src/shared/redux/types/settings.ts
@@ -1,0 +1,9 @@
+import { Record } from 'immutable';
+
+type NotificationType = 'granted' | 'denied' | 'default';
+
+export type SettingsState = Record<{
+  user: any;
+  system: any;
+  notificationPermission: NotificationType;
+}>;
